Move model-specific functions to pivot table models from tabular mixin
These functions fit better in pivot table models than in DB editor.
Now there is less jumping between the view and the models.

Re #2574
