+++ spinetoolbox/spine_db_editor/widgets/tabular_view_mixin.py
@@ -12,7 +12,7 @@
 
 """Contains TabularViewMixin class."""
 from contextlib import contextmanager
-from itertools import chain, product
+from itertools import chain
 from collections import namedtuple
 from PySide6.QtCore import QModelIndex, Qt, Slot, QTimer
 from PySide6.QtGui import QAction, QIcon, QActionGroup
@@ -259,258 +259,6 @@ class TabularViewMixin:
                 return item
             item = item.parent_item
         return None
-
-    @staticmethod
-    def _make_get_id(action):
-        """Returns a function to compute the db_map-id tuple of an item."""
-        return {"add": lambda db_map, x: (db_map, x["id"]), "remove": lambda db_map, x: None}[action]
-
-    def get_db_map_entities(self):
-        """Returns a dict mapping db maps to a list of dict entity items in the current class.
-
-        Returns:
-            dict
-        """
-        return {
-            db_map: self.db_mngr.get_items_by_field(db_map, "entity", "class_id", class_id)
-            for db_map, class_id in self.current_class_id.items()
-        }
-
-    def load_empty_element_data(self, db_map_class_entities=None):
-        """Returns a dict containing all possible entity elements in the current class.
-
-        Args:
-            db_map_class_entities (dict, optional)
-
-        Returns:
-            dict: Key is db_map-object_id tuple, value is None.
-        """
-        if db_map_class_entities is None:
-            db_map_class_entities = {}
-        if not self.first_current_entity_class["dimension_id_list"]:
-            return {}
-        data = {}
-        for db_map in self.db_maps:
-            element_id_lists = []
-            all_given_ids = set()
-            for db_map_dimension_id in self.current_dimension_id_list:
-                dim_id = db_map_dimension_id.get(db_map)
-                elements = self.db_mngr.get_items_by_field(db_map, "entity", "class_id", dim_id)
-                ids = {item["id"]: None for item in elements}
-                given_elements = db_map_class_entities.get(db_map, {}).get(dim_id)
-                if given_elements is not None:
-                    given_ids = {item["id"]: None for item in given_elements}
-                    ids.update(given_ids)
-                    all_given_ids.update(given_ids.keys())
-                element_id_lists.append(list(ids.keys()))
-            db_map_data = {
-                tuple((db_map, id_) for id_ in element_ids) + (db_map,): None
-                for element_ids in product(*element_id_lists)
-                if not all_given_ids or all_given_ids.intersection(element_ids)
-            }
-            data.update(db_map_data)
-        return data
-
-    def load_full_element_data(self, db_map_entities=None, action="add"):
-        """Returns a dict of entity elements in the current class.
-
-        Args:
-            db_map_entities (dict, optional): a mapping from database map to entities in the current entity class
-            action (str): 'add' or 'remove'
-
-        Returns:
-            dict: Key is db_map-object id tuple, value is relationship id.
-        """
-        if not self.first_current_entity_class.get("dimension_id_list", None):
-            return {}
-        if db_map_entities is None:
-            db_map_entities = self.get_db_map_entities()
-        get_id = self._make_get_id(action)
-        return {
-            tuple((db_map, id_) for id_ in ent["element_id_list"]) + (db_map,): get_id(db_map, ent)
-            for db_map, entities in db_map_entities.items()
-            for ent in entities
-        }
-
-    def load_relationship_data(self):
-        """Returns a dict that merges empty and full relationship data.
-
-        Returns:
-            dict: Key is object id tuple, value is True if a relationship exists, False otherwise.
-        """
-        data = self.load_empty_element_data()
-        data.update(self.load_full_element_data())
-        return data
-
-    def load_scenario_alternative_data(self, db_map_scenarios=None, db_map_alternatives=None):
-        """Returns a dict containing all scenario alternatives.
-
-        Returns:
-            dict: Key is db_map-id tuple, value is None or rank.
-        """
-        if db_map_scenarios is None:
-            db_map_scenarios = {db_map: self.db_mngr.get_items(db_map, "scenario") for db_map in self.db_maps}
-        if db_map_alternatives is None:
-            db_map_alternatives = {db_map: self.db_mngr.get_items(db_map, "alternative") for db_map in self.db_maps}
-        data = {}
-        for db_map in self.db_maps:
-            scenario_alternative_ranks = {
-                x["id"]: {
-                    alt_id: k + 1
-                    for k, alt_id in enumerate(self.db_mngr.get_scenario_alternative_id_list(db_map, x["id"]))
-                }
-                for x in db_map_scenarios.get(db_map, [])
-            }
-            alternative_ids = [x["id"] for x in db_map_alternatives.get(db_map, [])]
-            db_map_data = {
-                ((db_map, scen_id), (db_map, alt_id), db_map): alternative_ranks.get(alt_id)
-                for scen_id, alternative_ranks in scenario_alternative_ranks.items()
-                for alt_id in alternative_ids
-            }
-            data.update(db_map_data)
-        return data
-
-    def _get_db_map_parameter_value_or_def_ids(self, item_type):
-        """Returns a dict mapping db maps to a list of integer parameter (value or def) ids from the current class.
-
-        Args:
-            item_type (str): either "parameter_value" or "parameter_definition"
-
-        Returns:
-            dict
-        """
-        return {
-            db_map: [x["id"] for x in self.db_mngr.get_items_by_field(db_map, item_type, "entity_class_id", class_id)]
-            for db_map, class_id in self.current_class_id.items()
-        }
-
-    def _get_db_map_parameter_values_or_defs(self, item_type):
-        """Returns a dict mapping db maps to list of dict parameter (value or def) items from the current class.
-
-        Args:
-            item_type (str): either "parameter_value" or "parameter_definition"
-
-        Returns:
-            dict
-        """
-        db_map_ids = self._get_db_map_parameter_value_or_def_ids(item_type)
-        return {
-            db_map: [self.db_mngr.get_item(db_map, item_type, id_) for id_ in ids] for db_map, ids in db_map_ids.items()
-        }
-
-    def load_empty_parameter_value_data(
-        self, db_map_entities=None, db_map_parameter_ids=None, db_map_alternative_ids=None
-    ):
-        """Returns a dict containing all possible combinations of entities and parameters for the current class
-        in all db_maps.
-
-        Args:
-            db_map_entities (dict, optional): if given, only load data for these db maps and entities
-            db_map_parameter_ids (dict, optional): if given, only load data for these db maps and parameter definitions
-            db_map_alternative_ids (dict, optional): if given, only load data for these db maps and alternatives
-
-        Returns:
-            dict: Key is a tuple object_id, ..., parameter_id, value is None.
-        """
-        if db_map_entities is None:
-            db_map_entities = self.get_db_map_entities()
-        if db_map_parameter_ids is None:
-            db_map_parameter_ids = {
-                db_map: [(db_map, id_) for id_ in ids]
-                for db_map, ids in self._get_db_map_parameter_value_or_def_ids("parameter_definition").items()
-            }
-        if db_map_alternative_ids is None:
-            db_map_alternative_ids = {
-                db_map: [
-                    (db_map, id_) for a in self.db_mngr.get_items(db_map, "alternative") if (id_ := a["id"]) is not None
-                ]
-                for db_map in self.db_maps
-            }
-        db_map_entity_ids = {
-            db_map: [
-                id_tuple
-                for e in entities
-                if (id_tuple := tuple((db_map, id_) for id_ in e["element_id_list"] or (e["id"],)))
-            ]
-            for db_map, entities in db_map_entities.items()
-        }
-        if not any(db_map_entity_ids.values()) and (current_dimension_id_list := self.current_dimension_id_list):
-            db_map_entity_ids = {
-                db_map: [tuple((db_map, None) for _ in current_dimension_id_list)] for db_map in self.db_maps
-            }
-        if not any(db_map_parameter_ids.values()):
-            db_map_parameter_ids = {db_map: [(db_map, None)] for db_map in self.db_maps}
-        if not any(db_map_alternative_ids.values()):
-            db_map_alternative_ids = {db_map: [(db_map, None)] for db_map in self.db_maps}
-        return {
-            entity_id + (parameter_id, alt_id, db_map): None
-            for db_map in self.db_maps
-            for entity_id in db_map_entity_ids.get(db_map, [])
-            for parameter_id in db_map_parameter_ids.get(db_map, [])
-            for alt_id in db_map_alternative_ids.get(db_map, [])
-        }
-
-    def load_full_parameter_value_data(self, db_map_parameter_values=None, action="add"):
-        """Returns a dict of parameter values for the current class.
-
-        Args:
-            db_map_parameter_values (list, optional)
-            action (str)
-
-        Returns:
-            dict: Key is a tuple object_id, ..., parameter_id, value is the parameter_value.
-        """
-        if db_map_parameter_values is None:
-            db_map_parameter_values = self._get_db_map_parameter_values_or_defs("parameter_value")
-        get_id = self._make_get_id(action)
-        return {
-            tuple((db_map, id_) for id_ in x["element_id_list"] or (x["entity_id"],))
-            + ((db_map, x["parameter_id"]), (db_map, x["alternative_id"]), db_map): get_id(db_map, x)
-            for db_map, items in db_map_parameter_values.items()
-            for x in items
-        }
-
-    def _indexes(self, value):
-        if value is None:
-            return []
-        db_map, id_ = value
-        return self.db_mngr.get_value_indexes(db_map, "parameter_value", id_)
-
-    def load_empty_expanded_parameter_value_data(
-        self, db_map_entities=None, db_map_parameter_ids=None, db_map_alternative_ids=None
-    ):
-        """Makes a dict of expanded parameter values for the current class.
-
-        Args:
-            db_map_parameter_values (list, optional)
-            action (str)
-
-        Returns:
-            dict: mapping from unique value id tuple to value tuple
-        """
-        data = self.load_empty_parameter_value_data(db_map_entities, db_map_parameter_ids, db_map_alternative_ids)
-        return {
-            key[:-3] + ((None, index),) + key[-3:]: value
-            for key, value in data.items()
-            for index in self._indexes(value)
-        }
-
-    def load_full_expanded_parameter_value_data(self, db_map_parameter_values=None, action="add"):
-        """Makes a dict of expanded parameter values for the current class.
-
-        Args:
-            db_map_parameter_values (list, optional)
-            action (str)
-
-        Returns:
-            dict: mapping from unique value id tuple to value tuple
-        """
-        data = self.load_full_parameter_value_data(db_map_parameter_values, action)
-        return {
-            key[:-3] + ((None, index),) + key[-3:]: value
-            for key, value in data.items()
-            for index in self._indexes(value)
-        }
 
     def get_pivot_preferences(self):
         """Returns saved pivot preferences.
